requirements.txt
# Core ML libraries
torch>=2.0.0
pytorch-lightning>=2.0.0
pytorch-forecasting>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Geospatial libraries
h3>=3.7.0
geopandas>=0.12.0
folium>=0.14.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Hyperparameter optimization
optuna>=3.0.0

# Configuration management
PyYAML>=6.0
hydra-core>=1.3.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
geopy>=2.4.1
holidays
